vault/requirements.txt
# ===== 核心Web框架 =====
Flask==3.1.2

# ===== HTTP请求和异步 =====
requests==2.32.5
gunicorn==23.0.0
# ===== 数据库 =====
SQLAlchemy==2.0.44

# ===== 爬虫相关 =====

# ===== 工具库 =====
pillow==11.3.0
python-dotenv>=1.2.1
# ===== 工具库 =====
loguru>=0.7.3

# ===== 开发工具（可选） =====
